tests: Fix -Wsign-conversion warnings for random ints in gobject tests
There’s a painful inconsistency in the types of the
`g_{test_rand,random,rand}_int{,_range}()` functions, which vary
arbitrarily between `gint32` and `guint32`.

Unfortunately since those functions mention `int` explicitly in the name
(and then some of them return an `unsigned` integer), I don’t see a way
to make the APIs consistent without significant deprecations or
additions.

So, for the moment, to fix various `-Wsign-conversion` warnings, plaster
the tests with casts.


Helps: #3405

+++ gobject/tests/signal-handler.c
@@ -147,7 +147,7 @@ test_disconnect_many_random (void)
 
   for (i = 0; i < n_handlers; i++)
     {
-      j = g_test_rand_int_range (0, n_handlers);
+      j = (guint) g_test_rand_int_range (0, (int) n_handlers);
       id = handlers[i];
       handlers[i] = handlers[j];
       handlers[j] = id;
@@ -189,7 +189,7 @@ test_disconnect_2_signals (void)
 
   for (i = 0; i < n_handlers; i++)
     {
-      j = g_test_rand_int_range (0, n_handlers);
+      j = (guint) g_test_rand_int_range (0, (int) n_handlers);
       id = handlers[i];
       handlers[i] = handlers[j];
       handlers[j] = id;
@@ -240,7 +240,7 @@ test_disconnect_2_objects (void)
 
   for (i = 0; i < n_handlers; i++)
     {
-      j = g_test_rand_int_range (0, n_handlers);
+      j = (guint) g_test_rand_int_range (0, (int) n_handlers);
       id = handlers[i];
       handlers[i] = handlers[j];
       handlers[j] = id;
@@ -282,7 +282,7 @@ test_block_many (void)
 
   for (i = 0; i < n_handlers; i++)
     {
-      j = g_test_rand_int_range (0, n_handlers);
+      j = (guint) g_test_rand_int_range (0, (int) n_handlers);
       id = handlers[i];
       handlers[i] = handlers[j];
       handlers[j] = id;
